# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zeus MOB MDM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus];

export const DeviceStatus = {
  online: "online",
  offline: "offline",
  idle: "idle",
  busy: "busy",
} as const;

export interface Device {
  id: number;
  name: string;
  model: string;
  os: string;
  osVersion: string;
  ipAddress: string;
  status: DeviceStatus;
  batteryLevel: number;
  lastSeen: string;
  socketId?: string | null;
  createdAt: string;
}

export interface DeviceStats {
  total: number;
  online: number;
  offline: number;
  idle: number;
  busy: number;
}

export interface DeviceEvent {
  id: number;
  deviceId: number;
  deviceName: string;
  type: string;
  message: string;
  createdAt: string;
}

export interface Command {
  id: string;
  label: string;
  description: string;
  icon: string;
}

export interface SendCommandBody {
  deviceId: number;
  command: string;
}

export interface CommandResult {
  success: boolean;
  message: string;
}

export type ListEventsParams = {
  deviceId?: number;
  limit?: number;
};