# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zeus MOB MDM API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all devices
 */
export const ListDevicesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  model: zod.string(),
  os: zod.string(),
  osVersion: zod.string(),
  ipAddress: zod.string(),
  status: zod.enum(["online", "offline", "idle", "busy"]),
  batteryLevel: zod.number(),
  lastSeen: zod.coerce.date(),
  socketId: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListDevicesResponse = zod.array(ListDevicesResponseItem);

/**
 * @summary Get device by ID
 */
export const GetDeviceParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDeviceResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  model: zod.string(),
  os: zod.string(),
  osVersion: zod.string(),
  ipAddress: zod.string(),
  status: zod.enum(["online", "offline", "idle", "busy"]),
  batteryLevel: zod.number(),
  lastSeen: zod.coerce.date(),
  socketId: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Device statistics summary
 */
export const GetDeviceStatsResponse = zod.object({
  total: zod.number(),
  online: zod.number(),
  offline: zod.number(),
  idle: zod.number(),
  busy: zod.number(),
});

/**
 * @summary List recent events
 */
export const ListEventsQueryParams = zod.object({
  deviceId: zod.coerce.number().optional(),
  limit: zod.coerce.number().optional(),
});

export const ListEventsResponseItem = zod.object({
  id: zod.number(),
  deviceId: zod.number(),
  deviceName: zod.string(),
  type: zod.string(),
  message: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListEventsResponse = zod.array(ListEventsResponseItem);

/**
 * @summary List available commands
 */
export const ListCommandsResponseItem = zod.object({
  id: zod.string(),
  label: zod.string(),
  description: zod.string(),
  icon: zod.string(),
});
export const ListCommandsResponse = zod.array(ListCommandsResponseItem);

/**
 * @summary Send command to device
 */
export const SendCommandBody = zod.object({
  deviceId: zod.number(),
  command: zod.string(),
});

export const SendCommandResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});