# lib/api-zod/src/generated/types/commandResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zeus MOB MDM API
 * OpenAPI spec version: 0.1.0
 */

export interface CommandResult {
  success: boolean;
  message: string;
}