# lib/api-zod/src/generated/types/device.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zeus MOB MDM API
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceStatus } from "./deviceStatus";

export interface Device {
  id: number;
  name: string;
  model: string;
  os: string;
  osVersion: string;
  ipAddress: string;
  status: DeviceStatus;
  batteryLevel: number;
  lastSeen: Date;
  socketId?: string | null;
  createdAt: Date;
}