# lib/api-zod/src/generated/types/deviceEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zeus MOB MDM API
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceEvent {
  id: number;
  deviceId: number;
  deviceName: string;
  type: string;
  message: string;
  createdAt: Date;
}